backend/agrovera_microservices/contact_service/service/serializers.py
from rest_framework import serializers
from service.models import Contact

class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = ['name', 'email', 'message']
